+++ packages/components/index.js
@@ -1,1 +1,9 @@
-export Button from './Button';
+import Button from "./Button";
+import FormItemEmail from "./FormItemEmail";
+import FormItemPassword from "./FormItemPassword";
+
+export {
+  Button,
+  FormItemEmail,
+  FormItemPassword
+}