+++ packages/styles/font/index.js
@@ -1,0 +1,18 @@
+export const fontPrimary = '"Montserrat", sans-serif';
+
+export const font10 = '10px';
+export const font12 = '12px';
+export const font14 = '14px';
+export const font16 = '16px';
+export const font18 = '18px';
+export const font20 = '20px';
+
+export const fontWeight100 = 100;
+export const fontWeight200 = 200;
+export const fontWeight300 = 300;
+export const fontWeight400 = 400; // Normal
+export const fontWeight500 = 500;
+export const fontWeight600 = 600;
+export const fontWeight700 = 700;
+export const fontWeight800 = 800;
+export const fontWeight900 = 900;