+++ packages/styles/variables/index.js
@@ -1,0 +1,13 @@
+// Button
+const button = {
+    small      : '28px',
+    medium     : '36px',
+    large      : '48px',
+    extralarge : '60px',
+    width50    : '50%',
+    width75    : '75%',
+    width100   : '100%'
+}
+export {
+    button
+}